чуть более хорошее решение со стиялми
